Fix /W3 warnings in examples

+++ examples/audio/03-load-wav/load-wav.c
@@ -80,7 +80,7 @@ int SDL_AppIterate(void *appstate)
        We're being lazy here, but if there's less than the entire wav file left to play,
        just shove a whole copy of it into the queue, so we always have _tons_ of
        data queued for playback. */
-    if (SDL_GetAudioStreamAvailable(stream) < wav_data_len) {
+    if (SDL_GetAudioStreamAvailable(stream) < (int)wav_data_len) {
         /* feed more data to the stream. It will queue at the end, and trickle out as the hardware needs more data. */
         SDL_PutAudioStreamData(stream, wav_data, wav_data_len);
     }
